+++ tweakwcs/imalign.py
@@ -26,10 +26,18 @@ from . import __version__
 
 __author__ = 'Mihai Cara'
 
-__all__ = ['fit_wcs', 'align_wcs']
+__all__ = ['fit_wcs', 'align_wcs', 'NotEnoughCatalogs']
 
 log = logging.getLogger(__name__)
 log.setLevel(logging.DEBUG)
+
+
+class NotEnoughCatalogs(ValueError):
+    """
+    An exception class used to notify that an alignment routine does not have
+    enough input catalogs to perform alignment.
+    """
+    pass
 
 
 @deprecated_renamed_argument('tpwcs', 'corrector', since='0.8.0')
@@ -465,6 +473,14 @@ def align_wcs(wcscat, refcat=None, ref_tpwcs=None, enforce_user_order=True,
         ``refcat`` catalog, an expanded ``refcat`` with a combination of
         source positions from all input images.
 
+    Raises
+    ------
+    NotEnoughCatalogs
+        This exception is raised when there are not enough input catalogs
+        to perform alignment. For example, ``wcscat`` must be a list of at least
+        two correctors when ``refcat`` is `None`, or be a single corrector
+        when a refernce catalog is provided via ``refcat``.
+
     Notes
     -----
     **1. Weights:**
@@ -649,8 +665,10 @@ def align_wcs(wcscat, refcat=None, ref_tpwcs=None, enforce_user_order=True,
 
     # check that we have enough input images:
     if (refcat is None and len(wcs_gcat) < 2) or len(wcs_gcat) == 0:
-        raise ValueError("Too few input images (or groups of images) with "
-                         "non-empty catalogs.")
+        raise NotEnoughCatalogs(
+            "Too few input images (or groups of images) with "
+            "non-empty catalogs."
+        )
 
     # get the first image to be aligned and
     # create reference catalog if needed: